requirements.txt
pyttsx3
speechrecognition
opencv-python
torch
transformers
ultralytics
Pillow
requests
pyaudio
